auth/requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
exceptiongroup==1.1.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
invoke==2.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.0.0
pycparser==2.21
pydantic==1.10.8
PyJWT==2.7.0
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-dotenv==0.5.2
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.2
uvicorn==0.22.0
websockets==11.0.3
